refactored algorithms into their own package
